add new utilities for migration of existing content and better JSdocs
