+++ packages/extension-mathematics/src/extensions/InlineMath.ts
@@ -2,7 +2,23 @@ import { InputRule, mergeAttributes, Node } from '@tiptap/core'
 import type { Node as PMNode } from '@tiptap/pm/model'
 import katex from 'katex'
 
+/**
+ * Configuration options for the InlineMath extension.
+ */
 export type InlineMathOptions = {
+  /**
+   * Optional click handler for inline math nodes.
+   * Called when a user clicks on an inline math expression in the editor.
+   *
+   * @param node - The ProseMirror node representing the inline math element
+   * @param pos - The position of the node within the document
+   * @example
+   * ```ts
+   * onClick: (node, pos) => {
+   *   console.log('Inline math clicked:', node.attrs.latex, 'at position:', pos)
+   * }
+   * ```
+   */
   onClick?: (node: PMNode, pos: number) => void
 }
