+++ packages/extension-mathematics/src/extensions/index.ts
@@ -1,3 +1,2 @@
 export * from './BlockMath.js'
 export * from './InlineMath.js'
-export * from './MathMigration.js'
