+++ packages/extension-mathematics/src/index.ts
@@ -3,5 +3,6 @@ import { Mathematics } from './mathematics.js'
 export * from './extensions/index.js'
 export * from './mathematics.js'
 export * from './types.js'
+export * from './utils.js'
 
 export default Mathematics
